Add experimental support for Arduino 101
Add support for Arduino 101, by providing functions to switch between
read and write mode. For usage with external buffer with direction pin,
no change is needed.
Note that Arduino 101 use 3.3V (5V tolerant), so it is not compatible
with input buffers (74HC126D) of Dynamixel motors in theory.
It seems to work in practice, at least for low noise environment, but
external buffer/level converter is recommended.